refactor(guardian-ui): move all types into single directory for better sharing
