+++ apps/guardian-ui/src/components/SetConfiguration.tsx
@@ -14,12 +14,7 @@ import {
 import { useTranslation } from '@fedimint/utils';
 import { FormGroup, FormGroupHeading } from '@fedimint/ui';
 import { useSetupContext } from '../hooks';
-import {
-  BitcoinRpc,
-  ConfigGenParams,
-  GuardianRole,
-  Network,
-} from '../setup/types';
+import { BitcoinRpc, ConfigGenParams, GuardianRole, Network } from '../types';
 import { ReactComponent as FedimintLogo } from '../assets/svgs/fedimint.svg';
 import { ReactComponent as BitcoinLogo } from '../assets/svgs/bitcoin.svg';
 import { ReactComponent as ArrowRightIcon } from '../assets/svgs/arrow-right.svg';
